% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_countries.R
\name{carb_countries_ts}
\alias{carb_countries_ts}
\title{Time Series Plot}
\usage{
carb_countries_ts(
  countries_list,
  record_type,
  indicator,
  doforecast,
  yearforecast
)
}
\arguments{
\item{countries_list}{List of countries chosen by the user.}

\item{record_type}{Record chosen by the user.}

\item{doforecast}{Forecast.}

\item{yearforecast}{Number of year for which do the forecast.}
}
\value{
A Time Series plot with the chosen countries ecological footprint.
}
\description{
Plot Times Series of countries ecological footprints, to observe its evolution over time.
}
\examples{
country_list <- list("Spain", "Italy","France", "Germany","Switzerland")
carb_countries_ts(country_list, "EFProdTotGHA", quo(fishing_ground), "No", 10)
}
