% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{G6_stats}
\alias{G6_stats}
\title{Statistics}
\usage{
G6_stats(countries_list, record_type, indicator, year_range)
}
\arguments{
\item{countries_list}{List of countries chosen by the user.}

\item{record_type}{Record chosen by the user.}

\item{indicator}{Indicator selected by the user (crop land, grazing land, forest land, fishing ground, built-up land, carbon and total).}

\item{year_range}{Period over which the statistics are given.}
}
\value{
A table of statistics
}
\description{
The function returns a kable displaying statistics for selected country/ies, record type, production surface area's type and selected period. The statistics displayed are the minimum, median, mean, standard deviation, maximum, evolution and world proportion.
}
\examples{
G6_stats(c("France", "Switzerland", "World"), "AreaPerCap", "crop_land", (1990:2010))
}
