% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app}
\usage{
launch_app()
}
\value{
shiny app UI
}
\description{
The function calls the shiny app once the package is loaded
}
\examples{
launch_app
}
