% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_overview.R
\name{map_overview}
\alias{map_overview}
\title{Map by record type in 2016}
\usage{
map_overview(record_type)
}
\arguments{
\item{record_type}{Record type selected by the user ("AreaPerCap", "AreaTotHA", "BiocapPerCap", "BiocapTotGHA", "EFConsPerCap", "EFConsTotGHA", "EFProdPerCap", "EFProdTotGHA")}
}
\value{
An interactive World map
}
\description{
The function returns an interactive World map of available countries displaying the total ecological footprint (crop land, grazing land, forest land, fishing ground, built-up land and carbon) of different production sites, by selected record type. The user can move the mouse cursor on a country and see more detailed information on the different production sites' ecological footprint.
}
\examples{
G6::map_overview("AreaPerCap")
}
