% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot and stats.R
\name{sustainability}
\alias{sustainability}
\title{Ecological Footprint Plot}
\usage{
sustainability(countries_select)
}
\arguments{
\item{countries_select}{Countries user is selecting to observe their ecological footprint.}
}
\value{
A plot with the chosen countries ecological footprint.
}
\description{
Plot countries ecological footprints so that we can compare them.
}
\examples{
countries <- list("China","Germany","India")
sustainability(countries)
}
