% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_plot.R
\name{timeseries}
\alias{timeseries}
\title{Time Series Plot}
\usage{
timeseries(countries_list, record_type, indicator, doforecast, yearforecast)
}
\arguments{
\item{countries_list}{List of countries chosen by the user.}

\item{record_type}{Record chosen by the user.}

\item{indicator}{Indicator selected by the user (crop land, grazing land, forest land, fishing ground, built-up land, carbon and total).}

\item{doforecast}{Option to display forecasts or not.}

\item{yearforecast}{Number of years over which to forecast.}
}
\value{
A Time Series of the evolution of ecological footprint of chosen countries.
}
\description{
The function returns a times series of the evolution of different production surface areas' ecological footprint over the years for selected country/ies. The plot can display some forecasting if the user selects the option to do so.
}
\examples{
country_list <- list("Spain", "Italy","France", "Germany","Switzerland")
timeseries(country_list, "EFProdTotGHA", quo(fishing_ground), "No", 10)
}
