% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{worldmap}
\alias{worldmap}
\title{World Map}
\usage{
worldmap(country_selected, indicator, record_chosen)
}
\value{
A World Map with the countries ecological footprint.
}
\description{
Plot a world map displaying countries ecological footprintS.
}
