% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap_overview.R
\name{worldmap_overview}
\alias{worldmap_overview}
\title{Map by record type in 2016}
\usage{
worldmap_overview(record_type)
}
\arguments{
\item{record_type}{Record type selected by the user ("AreaPerCap", "AreaTotHA", "BiocapPerCap", "BiocapTotGHA", "EFConsPerCap", "EFConsTotGHA", "EFProdPerCap", "EFProdTotGHA")}
}
\value{
An interactive World map
}
\description{
The function returns an interactive World map of available countries displaying the total ecological footprint of different production sites (crop land, grazing land, forest land, fishing ground, built-up land and carbon), by selected record type ("AreaPerCap", "AreaTotHA", "BiocapPerCap", "BiocapTotGHA", "EFConsPerCap", "EFConsTotGHA", "EFProdPerCap", "EFProdTotGHA"). The user can move the mouse cursor on a country and see more detailed information on the different production sites' ecological footprint.
}
\examples{
worldmap_overview("AreaPerCap")
}
