% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G6_timeseries.R
\name{G6_timeseries}
\alias{G6_timeseries}
\title{Time Series Plot}
\usage{
G6_timeseries(countries_list, record_type, indicator)
}
\arguments{
\item{countries_list}{List of countries chosen by the user.}

\item{record_type}{Record chosen by the user.}

\item{indicator}{Indicator selected by the user (crop land, grazing land, forest land, fishing ground, built-up land, carbon and total).}
}
\value{
A Time Series of the evolution of ecological footprint of chosen countries.
}
\description{
The function returns times series of the evolution of different production surface areas' ecological footprints over the years for one or more selected countries.
}
\examples{
country_list <- list("Spain", "Italy","France", "Germany","Switzerland")
G6_timeseries(country_list, "EFProdTotGHA", "fishing_ground")
}
