% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_shiny.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app}
\usage{
launch_app()
}
\value{
shiny app UI
}
\description{
Function to start the shiny app once the package is loaded
}
\examples{
run_shiny()
}
